# Draft: ENH: Improve performance for np.result_type