Improve performance of  `np.result_type`

* Use `NPY_ALLOC_WORKSPACE` in `array_result_type`
* Avoid increfs/decrefs in `PyArray_ResultType`
* Fast path for array method lookup on dtypes
* Fast path for identical arguments to `np.result_type`

Benchmark:
```
result_type(x): Mean +- std dev: [main] 156 ns +- 8 ns -> [pr] 135 ns +- 2 ns: 1.16x faster
result_type(x, x): Mean +- std dev: [main] 195 ns +- 5 ns -> [pr] 150 ns +- 4 ns: 1.30x faster
result_type(x, x32): Mean +- std dev: [main] 207 ns +- 5 ns -> [pr] 174 ns +- 5 ns: 1.18x faster
result_type(x, x32, x): Mean +- std dev: [main] 224 ns +- 11 ns -> [pr] 191 ns +- 5 ns: 1.17x faster
result_type(d, d): Mean +- std dev: [main] 269 ns +- 12 ns -> [pr] 177 ns +- 7 ns: 1.52x faster
result_type(d, d32): Mean +- std dev: [main] 276 ns +- 10 ns -> [pr] 214 ns +- 6 ns: 1.29x faster
result_type(d, d32, x): Mean +- std dev: [main] 303 ns +- 11 ns -> [pr] 245 ns +- 9 ns: 1.23x faster

Geometric mean: 1.26x faster
```
<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->