@eendebakpt 2025-04-15
There are 2 failing tests due to a `warning: ‘arr’ may be used uninitialized [-Wmaybe-uninitialized]`. I think this is a false positive: the buffer `buf` is not initialized, but we are not using the buffer contents in the comparison (we are comparing the buffer address, which is initialized).