@eendebakpt 2025-04-15
It was not a false positive after all, the `arr` could be uninitialized for the zero argument case.